Updated the code injector to handle imports, data reader and canvas code injection in the notebook.
